Split User component in two, one expandable and one fixed
