+++ src/pagePartials/badge/preview/BadgeOwnerPreview.tsx
@@ -1,60 +1,15 @@
-import { useRouter } from 'next/router'
-import React, { useRef, useState } from 'react'
+import React from 'react'
 
-import AddRoundedIcon from '@mui/icons-material/AddRounded'
-import EmailOutlinedIcon from '@mui/icons-material/EmailOutlined'
-import RemoveRoundedIcon from '@mui/icons-material/RemoveRounded'
-import TwitterIcon from '@mui/icons-material/Twitter'
-import { Box, IconButton, Menu, MenuProps, Stack, Typography, alpha, styled } from '@mui/material'
-import { ButtonV2, IconDiscord, colors } from '@thebadge/ui-library'
-import { useTranslation } from 'next-export-i18n'
+import { colors } from '@thebadge/ui-library'
 
-import TBUserAvatar from '@/src/components/common/TBUserAvatar'
-import { Address } from '@/src/components/helpers/Address'
+import TBUserInfoExpandablePreview from '@/src/components/common/TBUserInfoMenuPreview'
 import { useUserById } from '@/src/hooks/subgraph/useUserById'
 import useS3Metadata from '@/src/hooks/useS3Metadata'
-import { truncateStringInTheMiddle } from '@/src/utils/strings'
 import { CreatorMetadata } from '@/types/badges/Creator'
 
-const Wrapper = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  flex: '1',
-  justifyContent: 'space-between',
-  border: `1px solid ${colors.purple}`,
-  filter: 'drop-shadow(0px 0px 8px rgba(255, 255, 255, 0.3))',
-  borderRadius: theme.spacing(1, 1, 0, 0),
-  padding: theme.spacing(0.1, 2, 0.1, 2),
-}))
-
-type OwnerDisplayProps = MenuProps & { width?: number }
-const OwnerDisplay = styled(Menu)<OwnerDisplayProps>(({ theme, width }) => ({
-  '& .MuiPaper-root': {
-    backgroundColor: '#000',
-    backgroundImage: 'none',
-    boxShadow: `0px 0px 4px ${alpha(colors.purple, 0.6)}`,
-    padding: theme.spacing(1, 2, 1, 2),
-    borderRadius: theme.spacing(0, 0, 1, 1),
-    width: `${width}px`,
-    maxWidth: `${width}px`,
-  },
-}))
-
 export default function BadgeOwnerPreview({ ownerAddress }: { ownerAddress: string }) {
-  const router = useRouter()
-  const { t } = useTranslation()
-
-  const [anchorEl, setAnchorEl] = useState<null | HTMLDivElement>(null)
-  const wrapperRef = useRef<HTMLDivElement>()
-  const open = Boolean(anchorEl)
-
   if (!ownerAddress) {
     throw `No ownerAddress provided`
-  }
-  const handleClick = () => {
-    if (wrapperRef?.current) setAnchorEl(wrapperRef.current)
-  }
-  const handleClose = () => {
-    setAnchorEl(null)
   }
 
   const owner = useUserById(ownerAddress)
@@ -62,71 +17,11 @@ export default function BadgeOwnerPreview({ ownerAddress }: { ownerAddress: stri
   const ownerMetadata = resMetadata.data?.content
 
   return (
-    <>
-      <Wrapper ref={wrapperRef}>
-        <Box alignItems="center" display="flex" gap={1}>
-          <Typography variant="dAppTitle5">{t('badge.viewBadge.owner.address')}</Typography>
-          <Address address={ownerAddress} showCopyButton={false} showExternalLink={false} />
-        </Box>
-
-        <IconButton onClick={handleClick} size="small">
-          {!open ? <AddRoundedIcon color="purple" /> : <RemoveRoundedIcon color="purple" />}
-        </IconButton>
-      </Wrapper>
-      <OwnerDisplay
-        anchorEl={anchorEl}
-        onClose={handleClose}
-        open={open}
-        width={wrapperRef.current?.getBoundingClientRect().width}
-      >
-        <Box display="flex" flex="1" gap={2}>
-          <TBUserAvatar
-            address={ownerAddress}
-            isVerified={owner.data?.isVerified}
-            src={ownerMetadata?.logo?.s3Url}
-            sx={{ border: '1px solid white' }}
-          />
-          <Stack gap={1}>
-            <Typography variant="dAppHeadline2">
-              {ownerMetadata?.name || truncateStringInTheMiddle(ownerAddress, 8, 6)}
-            </Typography>
-            <Stack flex="1" justifyContent="space-evenly">
-              {ownerMetadata?.email && (
-                <Typography variant="dAppTitle2">
-                  <EmailOutlinedIcon sx={{ mr: 1 }} />
-                  {ownerMetadata?.email}
-                </Typography>
-              )}
-              {ownerMetadata?.twitter && (
-                <Typography variant="dAppTitle2">
-                  <TwitterIcon sx={{ mr: 1 }} />
-                  {ownerMetadata?.twitter}
-                </Typography>
-              )}
-              {ownerMetadata?.discord && (
-                <Typography variant="dAppTitle2">
-                  <IconDiscord sx={{ mr: 1 }} />
-                  {ownerMetadata?.discord}
-                </Typography>
-              )}
-            </Stack>
-            <ButtonV2
-              backgroundColor={colors.transparent}
-              fontColor={colors.purple}
-              onClick={() => router.push(`/profile/${ownerAddress}`)}
-              sx={{
-                borderRadius: 2,
-                mt: 2,
-              }}
-              variant="outlined"
-            >
-              <Typography fontSize="12px !important">
-                {t('badge.viewBadge.owner.profile')}
-              </Typography>
-            </ButtonV2>
-          </Stack>
-        </Box>
-      </OwnerDisplay>
-    </>
+    <TBUserInfoExpandablePreview
+      color={colors.purple}
+      isVerified={owner?.data?.isVerified}
+      metadata={ownerMetadata}
+      userAddress={ownerAddress}
+    />
   )
 }
